Add issue logic for selecting other issues
